Implement new design for learningpath save page when shared
