+++ src/containers/MyNdla/Learningpath/components/LearningpathActionHooks.tsx
@@ -99,13 +99,7 @@ export const useLearningpathActionHooks = (learningpath?: GQLMyNdlaLearningpathF
       text: t("myNdla.learningpath.menu.share"),
       value: "shareLearningPath",
       icon: <ShareLine />,
-      modalContent: (close) => (
-        <LearningpathShareDialogContent
-          learningpath={learningpath}
-          onClose={close}
-          onCopyText={() => copyLearningpathSharingLink(learningpath.id)}
-        />
-      ),
+      modalContent: (close) => <LearningpathShareDialogContent learningpath={learningpath} onClose={close} />,
       onClick: !isShared
         ? async () => {
             const res = await updateLearningpathStatus({
